+++ packages/nodes-base/nodes/Google/BigQuery/v2/transport/index.ts
@@ -1,58 +1,9 @@
 import type { OptionsWithUri } from 'request';
-import moment from 'moment-timezone';
-import * as jwt from 'jsonwebtoken';
 
 import type { IExecuteFunctions, IExecuteSingleFunctions, ILoadOptionsFunctions } from 'n8n-core';
 import type { IDataObject, JsonObject } from 'n8n-workflow';
 import { NodeApiError, NodeOperationError } from 'n8n-workflow';
-
-async function getAccessToken(
-	this: IExecuteFunctions | IExecuteSingleFunctions | ILoadOptionsFunctions,
-	credentials: IDataObject,
-): Promise<IDataObject> {
-	//https://developers.google.com/identity/protocols/oauth2/service-account#httprest
-
-	const privateKey = (credentials.privateKey as string).replace(/\\n/g, '\n').trim();
-
-	const scopes = ['https://www.googleapis.com/auth/bigquery'];
-
-	const now = moment().unix();
-
-	const signature = jwt.sign(
-		{
-			iss: credentials.email as string,
-			sub: credentials.delegatedEmail || (credentials.email as string),
-			scope: scopes.join(' '),
-			aud: 'https://oauth2.googleapis.com/token',
-			iat: now,
-			exp: now + 3600,
-		},
-		privateKey,
-		{
-			algorithm: 'RS256',
-			header: {
-				kid: privateKey,
-				typ: 'JWT',
-				alg: 'RS256',
-			},
-		},
-	);
-
-	const options: OptionsWithUri = {
-		headers: {
-			'Content-Type': 'application/x-www-form-urlencoded',
-		},
-		method: 'POST',
-		form: {
-			grant_type: 'urn:ietf:params:oauth:grant-type:jwt-bearer',
-			assertion: signature,
-		},
-		uri: 'https://oauth2.googleapis.com/token',
-		json: true,
-	};
-
-	return this.helpers.request(options);
-}
+import { getGoogleAccessToken } from '../../../GenericFunctions';
 
 export async function googleApiRequest(
 	this: IExecuteFunctions | IExecuteSingleFunctions | ILoadOptionsFunctions,
@@ -94,7 +45,9 @@ export async function googleApiRequest(
 				throw new NodeOperationError(this.getNode(), 'No credentials got returned!');
 			}
 
-			const { access_token } = await getAccessToken.call(this, credentials as IDataObject);
+			const scopes = ['https://www.googleapis.com/auth/bigquery'];
+
+			const { access_token } = await getGoogleAccessToken.call(this, credentials, scopes);
 
 			options.headers!.Authorization = `Bearer ${access_token}`;
 			return await this.helpers.request(options);
