+++ packages/nodes-base/nodes/Google/Gmail/GenericFunctions.ts
@@ -4,7 +4,6 @@ import { simpleParser } from 'mailparser';
 
 import type {
 	IBinaryKeyData,
-	ICredentialDataDecryptedObject,
 	IDataObject,
 	IExecuteFunctions,
 	IExecuteSingleFunctions,
@@ -15,10 +14,6 @@ import type {
 	JsonObject,
 } from 'n8n-workflow';
 import { NodeApiError, NodeOperationError } from 'n8n-workflow';
-
-import moment from 'moment-timezone';
-
-import jwt from 'jsonwebtoken';
 
 import { DateTime } from 'luxon';
 
@@ -44,62 +39,7 @@ export interface IAttachments {
 }
 
 import MailComposer from 'nodemailer/lib/mail-composer';
-
-async function getAccessToken(
-	this: IExecuteFunctions | IExecuteSingleFunctions | ILoadOptionsFunctions | IPollFunctions,
-	credentials: ICredentialDataDecryptedObject,
-): Promise<IDataObject> {
-	//https://developers.google.com/identity/protocols/oauth2/service-account#httprest
-
-	const scopes = [
-		'https://www.googleapis.com/auth/gmail.labels',
-		'https://www.googleapis.com/auth/gmail.addons.current.action.compose',
-		'https://www.googleapis.com/auth/gmail.addons.current.message.action',
-		'https://mail.google.com/',
-		'https://www.googleapis.com/auth/gmail.modify',
-		'https://www.googleapis.com/auth/gmail.compose',
-	];
-
-	const now = moment().unix();
-
-	credentials.email = (credentials.email as string).trim();
-	const privateKey = (credentials.privateKey as string).replace(/\\n/g, '\n').trim();
-
-	const signature = jwt.sign(
-		{
-			iss: credentials.email,
-			sub: credentials.delegatedEmail || credentials.email,
-			scope: scopes.join(' '),
-			aud: 'https://oauth2.googleapis.com/token',
-			iat: now,
-			exp: now + 3600,
-		},
-		privateKey,
-		{
-			algorithm: 'RS256',
-			header: {
-				kid: privateKey,
-				typ: 'JWT',
-				alg: 'RS256',
-			},
-		},
-	);
-
-	const options: OptionsWithUri = {
-		headers: {
-			'Content-Type': 'application/x-www-form-urlencoded',
-		},
-		method: 'POST',
-		form: {
-			grant_type: 'urn:ietf:params:oauth:grant-type:jwt-bearer',
-			assertion: signature,
-		},
-		uri: 'https://oauth2.googleapis.com/token',
-		json: true,
-	};
-
-	return this.helpers.request(options);
-}
+import { getGoogleAccessToken } from '../GenericFunctions';
 
 export async function googleApiRequest(
 	this: IExecuteFunctions | IExecuteSingleFunctions | ILoadOptionsFunctions | IPollFunctions,
@@ -139,7 +79,16 @@ export async function googleApiRequest(
 			const credentials = await this.getCredentials('googleApi');
 			credentialType = 'googleApi';
 
-			const { access_token } = await getAccessToken.call(this, credentials);
+			const scopes = [
+				'https://www.googleapis.com/auth/gmail.labels',
+				'https://www.googleapis.com/auth/gmail.addons.current.action.compose',
+				'https://www.googleapis.com/auth/gmail.addons.current.message.action',
+				'https://mail.google.com/',
+				'https://www.googleapis.com/auth/gmail.modify',
+				'https://www.googleapis.com/auth/gmail.compose',
+			];
+
+			const { access_token } = await getGoogleAccessToken.call(this, credentials, scopes);
 
 			(options.headers as IDataObject).Authorization = `Bearer ${access_token}`;
 		}
